Reuse util methods for all batch job
