+++ java/openmldb-batchjob/src/main/scala/com/_4paradigm/openmldb/batchjob/ImportOnlineData.scala
@@ -16,27 +16,19 @@
 
 package com._4paradigm.openmldb.batchjob
 
-import com._4paradigm.openmldb.batch.api.OpenmldbSession
-import org.apache.spark.SparkFiles
+import com._4paradigm.openmldb.batchjob.util.OpenmldbJobUtil
 import org.apache.spark.sql.SparkSession
 
 object ImportOnlineData {
 
   def main(args: Array[String]): Unit = {
-    if (args.length < 1) {
-      throw new Exception(s"Require args: sql but get args: ${args.mkString(",")}")
-    }
-
+    OpenmldbJobUtil.checkOneSqlArgument(args)
     importOnlineData(args(0))
   }
 
   def importOnlineData(sqlFilePath: String): Unit = {
-    val sess = new OpenmldbSession(SparkSession.builder().config("openmldb.loaddata.mode", "online").getOrCreate())
-
-    val sqlText = scala.io.Source.fromFile(SparkFiles.get(sqlFilePath)).mkString
-    sess.sql(sqlText)
-
-    sess.close()
+    val spark = SparkSession.builder().config("openmldb.loaddata.mode", "online").getOrCreate()
+    OpenmldbJobUtil.runOpenmldbSql(spark, sqlFilePath)
   }
 
 }
