+++ java/openmldb-batchjob/src/main/scala/com/_4paradigm/openmldb/batchjob/RunBatchAndShow.scala
@@ -17,39 +17,22 @@
 package com._4paradigm.openmldb.batchjob
 
 import com._4paradigm.openmldb.batch.api.OpenmldbSession
-import org.apache.spark.SparkFiles
+import com._4paradigm.openmldb.batchjob.util.OpenmldbJobUtil
 import org.apache.spark.sql.SparkSession
-import scala.reflect.io.File
 
 object RunBatchAndShow {
 
   def main(args: Array[String]): Unit = {
-    if (args.length < 1) {
-      throw new Exception(s"Require args: sql but get args: ${args.mkString(",")}")
-    }
-
+    OpenmldbJobUtil.checkOneSqlArgument(args)
     runBatchSql(args(0))
   }
 
   def runBatchSql(sqlFilePath: String): Unit = {
-    val sess = new OpenmldbSession(SparkSession.builder().getOrCreate())
+    val spark = SparkSession.builder().getOrCreate()
+    val sqlText = OpenmldbJobUtil.getSqlFromFile(spark, sqlFilePath)
 
-    val sparkMaster = sess.getSparkSession.conf.get("spark.master")
-
-    val actualSqlFilePath = if (sparkMaster.equals("local")) {
-      SparkFiles.get(sqlFilePath)
-    } else {
-      sqlFilePath
-    }
-
-    if (!File(actualSqlFilePath).exists) {
-      throw new Exception("SQL file does not exist in " + actualSqlFilePath)
-    }
-
-    val sqlText = scala.io.Source.fromFile(actualSqlFilePath).mkString
-
+    val sess = new OpenmldbSession(spark)
     sess.sql(sqlText).show()
-
     sess.close()
   }
